integrate ibis relation into existing code
